+++ package.json
@@ -21,7 +21,7 @@
     "url": "https://github.com/palmetto/palmetto-components"
   },
   "dependencies": {
-    "@palmetto/palmetto-design-tokens": "0.44.5",
+    "@palmetto/palmetto-design-tokens": "0.46.0",
     "@popperjs/core": "^2.5.3",
     "@reach/dialog": "^0.11.2",
     "classnames": "^2.2.6",
