+++ src/components/Toast/Toast.Overview.stories.mdx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { action } from '@storybook/addon-actions';
 import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs/blocks';
-import { ToastContainer, toast } from './ToastContainer';
+import { ToastContainer, toast } from './';
 import { Box } from '../Box/Box';
 import { Button } from '../Button/Button';
 
@@ -15,13 +15,12 @@ import { Button } from '../Button/Button';
 
 # Toast
 
-
 <Canvas isExpanded>
   <Story name="Default">
     {() => {
       const handleClick = () => {
         toast('hey oh!', {
-          position: 'bottom-right',
+          position: 'top-center',
         });
       }
       return (
@@ -38,4 +37,4 @@ import { Button } from '../Button/Button';
 
 ## Props
 
-<ArgsTable of={OptionTile} />
+<ArgsTable of={ToastContainer} />
