+++ src/components/Toast/ToastNotification.module.scss
@@ -65,3 +65,11 @@
     opacity: 0;
   }
 }
+
+.toast-notification-enter-top { animation: toastEnterAnimationTop 0.35s cubic-bezier(0.21, 1.02, 0.73, 1) forwards; }
+.toast-notification-exit-top { animation: toastExitAnimationTop 0.4s forwards cubic-bezier(0.06, 0.71, 0.55, 1); }
+.toast-notification-enter-bottom { animation: toastEnterAnimationBottom 0.35s cubic-bezier(0.21, 1.02, 0.73, 1) forwards; }
+.toast-notification-exit-bottom { animation: toastExitAnimationBottom 0.4s forwards cubic-bezier(0.06, 0.71, 0.55, 1); }
+.toast-notification-fade-in { animation: fadeIn 0.35s cubic-bezier(0.21, 1.02, 0.73, 1) forwards; }
+.toast-notification-fade-out { animation: fadeOut 0.4s forwards cubic-bezier(0.06, 0.71, 0.55, 1); }
+
