+++ src/lib/isFunction.ts
@@ -1,0 +1,7 @@
+
+import { ValueOrFunction, ValueFunction } from '../types';
+
+export const isFunction = <TValue, TArg>(
+  valOrFunction: ValueOrFunction<TValue, TArg>
+): valOrFunction is ValueFunction<TValue, TArg> =>
+  typeof valOrFunction === 'function';
