+++ src/lib/prefersReducedMotion.ts
@@ -1,0 +1,11 @@
+export const prefersReducedMotion = (() => {
+  let shouldReduceMotion: boolean | undefined = undefined;
+
+  return () => {
+    if (shouldReduceMotion === undefined) {
+      const mediaQuery = window?.matchMedia('(prefers-reduced-motion: reduce)');
+      shouldReduceMotion = !mediaQuery || mediaQuery.matches;
+    }
+    return shouldReduceMotion;
+  };
+})();
