+++ src/lib/resolveValue.ts
@@ -1,0 +1,7 @@
+import { ValueOrFunction } from '../types';
+import { isFunction } from './isFunction';
+
+export const resolveValue = <TValue, TArg>(
+  valOrFunction: ValueOrFunction<TValue, TArg>,
+  arg: TArg
+): TValue => (isFunction(valOrFunction) ? valOrFunction(arg) : valOrFunction);
